requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.3
django-crispy-forms==1.9.1
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
